The 'nflverse' - A set of packages dedicated to data of the National Football League.

The 'nflverse' is a set of packages dedicated to data of the National Football League. This package is designed to make it easy to install and load multiple 'nflverse' packages in a single step. Learn more about the 'nflverse' at https://nflverse.nflverse.com/.

See also

Author

Maintainer: Sebastian Carl mrcaseb@gmail.com

Authors:

Ben Baldwin bbaldwin206@gmail.com

Lee Sharpe

Tan Ho tan@tanho.ca (ORCID)

John Edwards edwards1860@gmail.com (ORCID)