Get a Situation Report on System, nflverse/ffverse Package Versions and Dependencies

This function gives a quick overview of the versions of R and the operating system as well as the versions of nflverse/ffverse packages and their dependencies. It's primarily designed to help you get a quick idea of what's going on when you're helping someone else debug a problem.

nflverse_sitrep(
  pkg = c("nflreadr", "nflfastR", "nflseedR", "nfl4th", "nflplotR"),
  recursive = TRUE
)

Arguments

pkg

a character vector naming installed packages, or NULL (the default) meaning all nflverse packages. The function checks internally if all packages are installed and informs if that is not the case.

recursive

a logical indicating whether dependencies of pkg and their dependencies (and so on) should be included. Can also be a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

Examples

try(
nflverse_sitrep()
)
#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.1.2 (2021-11-01)   • Running under: macOS Big Sur 10.16
#> ── nflverse Packages ───────────────────────────────────────────────────────────
#> • nflreadr (1.1.2)  • nflseedR (1.0.2)  • nflplotR (1.0.0)  
#> • nflfastR (4.3.0)  • nfl4th   (1.0.1)    
#> ── nflverse Dependencies ───────────────────────────────────────────────────────
#> • MASS          (7.3-54)  • future     (1.23.0)   • pkgconfig   (2.0.3)    
#> • Matrix        (1.3-4)   • generics   (0.1.1)    • prettyunits (1.1.1)    
#> • R6            (2.5.1)   • ggplot2    (3.3.5)    • progress    (1.2.2)    
#> • RApiSerialize (0.1.0)   • globals    (0.14.0)   • progressr   (0.10.0)   
#> • RColorBrewer  (1.1-2)   • glue       (1.6.0)    • proto       (1.0.0)    
#> • Rcpp          (1.0.8)   • gsubfn     (0.7)      • purrr       (0.3.4)    
#> • RcppParallel  (5.1.5)   • gtable     (0.3.0)    • qs          (0.25.2)   
#> • askpass       (1.1)     • hms        (1.1.1)    • rappdirs    (0.3.3)    
#> • bit           (4.0.4)   • httr       (1.4.2)    • readr       (2.1.1)    
#> • bit64         (4.0.5)   • isoband    (0.2.5)    • rlang       (0.4.12)   
#> • cachem        (1.0.6)   • janitor    (2.1.0)    • scales      (1.1.1)    
#> • cli           (3.1.1)   • jsonlite   (1.7.3)    • snakecase   (0.11.0)   
#> • clipr         (0.7.1)   • labeling   (0.4.2)    • stringfish  (0.15.5)   
#> • codetools     (0.2-18)  • lattice    (0.20-45)  • stringi     (1.7.6)    
#> • colorspace    (2.0-2)   • lifecycle  (1.0.1)    • stringr     (1.4.0)    
#> • cpp11         (0.4.2)   • listenv    (0.8.0)    • sys         (3.4)      
#> • crayon        (1.4.2)   • lubridate  (1.8.0)    • tibble      (3.1.6)    
#> • curl          (4.3.2)   • magick     (2.7.3)    • tidyr       (1.1.4)    
#> • data.table    (1.14.2)  • magrittr   (2.0.1)    • tidyselect  (1.1.1)    
#> • digest        (0.6.29)  • memoise    (2.0.1)    • tzdb        (0.2.0)    
#> • dplyr         (1.0.7)   • mgcv       (1.8-38)   • utf8        (1.2.2)    
#> • ellipsis      (0.3.2)   • mime       (0.12)     • vctrs       (0.3.8)    
#> • fansi         (1.0.2)   • munsell    (0.5.0)    • viridisLite (0.4.0)    
#> • farver        (2.1.0)   • nlme       (3.1-153)  • vroom       (1.5.7)    
#> • fastmap       (1.1.0)   • openssl    (1.4.6)    • withr       (2.4.3)    
#> • fastrmodels   (1.0.2)   • parallelly (1.30.0)   • xgboost     (1.5.0.2)  
#> • furrr         (0.2.3)   • pillar     (1.6.4)      
#> ────────────────────────────────────────────────────────────────────────────────