Update nflverse Packages

Check all nflverse packages for available cran updates.

Usage

nflverse_update(recursive = FALSE, repos = getOption("repos"), devel = FALSE)

Arguments

recursive

If TRUE, will also list all strong dependencies of nflverse packages.

repos

the repositories to use to check for updates. Defaults to getOptions("repos").

devel

If TRUE the function will look for prebuilt development versions on https://nflverse.r-universe.dev/ so it is possible to install development versions without GitHub.

Value

Returns NULL invisibly. Called for side effects.

Examples

# \donttest{
try(
nflverse_update()
)
#> ✔ All nflverse packages are up-to-date!
try(
nflverse_update(devel = TRUE)
)
#> ℹ The following packages are out of date:
#> • nfl4th   (1.0.1 -> 1.0.1.9000)
#> • nflfastR (4.3.0 -> 4.3.0.9017)
#> • nflplotR (1.0.1 -> 1.0.1.9002)
#> • nflreadr (1.2.0 -> 1.2.0.14  )
#> ── Start a clean R session then run ────────────────────────────────────────────
#> options(repos = c(
#>         nflverse = "https://nflverse.r-universe.dev",
#>         getOption("repos")
#>       ))
#> install.packages(c("nfl4th", "nflfastR", "nflplotR", "nflreadr"))
#> ────────────────────────────────────────────────────────────────────────────────
# }